feat(ion-identity-client): rename user related classes and methods
