+++ packages/ion_identity_client/example/lib/pages/get_ion_connect_indexers/providers/ion_connect_indexers_notifier.dart
@@ -2,19 +2,18 @@
 
 import 'package:ion_client_example/providers/current_username_notifier.dart';
 import 'package:ion_client_example/providers/ion_client_provider.dart';
-import 'package:ion_identity_client/ion_client.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-part 'user_details_notifier.g.dart';
+part 'ion_connect_indexers_notifier.g.dart';
 
 @riverpod
-class UserDetailsNotifier extends _$UserDetailsNotifier {
+class IonConnectIndexersNotifier extends _$IonConnectIndexersNotifier {
   @override
-  FutureOr<UserDetails?> build() async {
+  FutureOr<List<String>?> build() async {
     return null;
   }
 
-  Future<void> fetchUserDetails(String userId) async {
+  Future<void> fetchIonConnectIndexers(String userId) async {
     state = const AsyncLoading();
 
     state = await AsyncValue.guard(() async {
@@ -24,7 +23,9 @@ class UserDetailsNotifier extends _$UserDetailsNotifier {
       }
 
       final ionIdentity = await ref.watch(ionClientProvider.future);
-      return await ionIdentity(username: currentUser).users.details(userId: userId);
+      return await ionIdentity(username: currentUser).users.ionConnectIndexers(
+            userId: userId,
+          );
     });
   }
 }
