+++ packages/ion_identity_client/lib/ion_client.dart
@@ -9,9 +9,9 @@ export 'src/core/types/ion_exception.dart';
 export 'src/core/types/user_token.dart';
 export 'src/ion_api_client.dart';
 export 'src/ion_client_config.dart';
-export 'src/users/get_user_connect_indexers/models/get_user_connect_indexers_response.dart';
-export 'src/users/get_user_details/models/user_details.dart';
-export 'src/users/set_user_connect_relays/models/set_user_connect_relays_response.dart';
+export 'src/users/ion_connect_indexers/models/ion_connect_indexers_response.dart';
+export 'src/users/set_ion_connect_relays/models/set_ion_connect_relays_response.dart';
+export 'src/users/user_details/models/user_details.dart';
 export 'src/wallets/models/wallet.dart';
 export 'src/wallets/models/wallet_signing_key.dart';
 export 'src/wallets/services/get_wallet_assets/exceptions/get_wallet_assets_exception.dart';
