+++ packages/ion_identity_client/lib/src/users/ion_users.dart
@@ -1,36 +1,36 @@
 // SPDX-License-Identifier: ice License 1.0
 
 import 'package:ion_identity_client/ion_client.dart';
-import 'package:ion_identity_client/src/users/get_user_connect_indexers/get_user_connect_indexers_service.dart';
-import 'package:ion_identity_client/src/users/get_user_details/get_user_details_service.dart';
-import 'package:ion_identity_client/src/users/set_user_connect_relays/set_user_connect_relays_service.dart';
+import 'package:ion_identity_client/src/users/ion_connect_indexers/get_user_connect_indexers_service.dart';
+import 'package:ion_identity_client/src/users/set_ion_connect_relays/set_ion_connect_relays_service.dart';
+import 'package:ion_identity_client/src/users/user_details/user_details_service.dart';
 
 class IONUsersClient {
   IONUsersClient(
     this._getUserDetailsService,
-    this._getUserConnectIndexersService,
-    this._setUserConnectRelaysService,
+    this._ionConnectIndexersService,
+    this._setIonConnectRelaysService,
   );
 
-  final GetUserDetailsService _getUserDetailsService;
-  final GetUserConnectIndexersService _getUserConnectIndexersService;
-  final SetUserConnectRelaysService _setUserConnectRelaysService;
+  final UserDetailsService _getUserDetailsService;
+  final IonConnectIndexersService _ionConnectIndexersService;
+  final SetIonConnectRelaysService _setIonConnectRelaysService;
 
-  Future<UserDetails> getUserDetails({
+  Future<UserDetails> details({
     required String userId,
   }) async =>
-      _getUserDetailsService.getUserDetails(userId: userId);
+      _getUserDetailsService.details(userId: userId);
 
-  Future<List<String>> getUserConnectIndexers({
+  Future<List<String>> ionConnectIndexers({
     required String userId,
   }) async =>
-      _getUserConnectIndexersService.getUserConnectIndexers(userId: userId);
+      _ionConnectIndexersService.ionConnectIndexers(userId: userId);
 
-  Future<SetUserConnectRelaysResponse> setUserConnectRelays({
+  Future<SetIonConnectRelaysResponse> setIonConnectRelays({
     required String userId,
     required List<String> followeeList,
   }) async =>
-      _setUserConnectRelaysService.setUserConnectRelays(
+      _setIonConnectRelaysService.setIonConnectRelays(
         userId: userId,
         followeeList: followeeList,
       );
