+++ packages/ion_identity_client/lib/src/users/user_details/data_sources/user_details_data_source.dart
@@ -1,0 +1,39 @@
+// SPDX-License-Identifier: ice License 1.0
+
+import 'package:ion_identity_client/ion_client.dart';
+import 'package:ion_identity_client/src/core/network/network_client.dart';
+import 'package:ion_identity_client/src/core/token_storage/token_storage.dart';
+import 'package:ion_identity_client/src/core/types/request_headers.dart';
+
+class UserDetailsDataSource {
+  UserDetailsDataSource(
+    this._networkClient,
+    this._tokenStorage,
+  );
+
+  final NetworkClient _networkClient;
+  final TokenStorage _tokenStorage;
+
+  static const usersDetailsPath = '/auth/users';
+
+  Future<UserDetails> fetchDetails({
+    required String username,
+    required String userId,
+  }) async {
+    final token = _tokenStorage.getToken(username: username);
+    if (token == null) {
+      throw const UnauthenticatedException();
+    }
+
+    final response = await _networkClient.get(
+      '$usersDetailsPath/$userId',
+      headers: RequestHeaders.getAuthorizationHeaders(
+        token: token.token,
+        username: username,
+      ),
+      decoder: UserDetails.fromJson,
+    );
+
+    return response;
+  }
+}
